Initial commit: just packaging the binary download
It is a kinda strange binary that looks for the
Resources folder in the parent directory. I also
skipped the "auto-run 32/64 bit" shim since it
didn't seem to do too well as a link-target.

To update this, it should be as simple as
replacing the /opt/FreeFileSync contents
with the unzip of a new version.